Added fix to Blob Checkpointer files, tests and updated changelog

+++ sdk/storage/Azure.Storage.DataMovement.Files.Shares/src/ShareFileDestinationCheckpointDetails.cs
@@ -487,7 +487,7 @@ namespace Azure.Storage.DataMovement.Files.Shares
             {
                 throw Storage.Errors.InvalidCheckpointNegativeOffset(offset);
             }
-            if (offset + length > streamLength)
+            if ((long)offset + length > streamLength)
             {
                 throw Storage.Errors.InvalidCheckpointOffsetLength(offset, length, streamLength);
             }
